+++ flutter/packages/x_color/CHANGELOG.md
@@ -1,3 +1,7 @@
 ## [0.0.1] - TODO: Add release date.
-
-* TODO: Describe initial release.
+## [0.0.2] - first implementation
+supports xcolor that accepts uri as the constructor input
+```dart
+var color = XColor("#FF23232");
+var icon = Icon(Icons.add, color: color);
+```